Create Queue.java
Manipulate a queue according to the given insert and remove commands and then output the string that is in the middle of the queue. If there is an even number of strings in the queue, thus two middle strings, output the one which is nearest the front. If a remove command is issued for an empty queue then nothing should happen.​

​

Sample Input​

INSERT this​

INSERT is​

INSERT how​

INSERT to​

REMOVE​

INSERT do​

REMOVE​

INSERT it​

​

Sample Output​

to