Create Co-Prime.java
Two integers are coprime if the only positive integer that divides into both of them is 1. In other words, the greatest common divisor of two coprime numbers is 1.​

If two numbers x and y are selected randomly, the odds that they will be coprime with e ach other is 61% (6 divided by PI^2). But what if we put some restrictions on the two numbers? For example, if x is a randomly selected number that is divisible by 7 and y is a randomly selected number that is divisible by 13? In this case, the odds that they are coprime is 49%.​​

In this question you are given a divisor of the first number and a divisor of the second number. You must output an integer from 0 to 100, which is the percentage probability that the two randomly selected numbers are coprime.