Create sortedStringArray.java
In this task the input is two numbers, followed by a series of strings. Those strings must be sorted and then one particular string must be outputted. The first number denotes the quantity of strings to follow. The second number indicates which string should be outputted following the sort. For example, if it's a 3, then output the string that falls in third place. If it's a 5 then output the string that falls in fifth place.​

The sorting that should be applied is as follows: sort words by their length, with the shortest words coming first. If two words have the same length, then sort them in reverse alphabetical order (i.e. with "zoo" coming ahead of "ape").